requirements.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
torch-geometric==2.3.1
numpy==1.24.4
matplotlib==3.7.1
scipy==1.10.1
networkx==3.1
tqdm==4.65.0
scikit-learn==1.3.0
pyyaml==6.0